python/rag/requirements.txt
acres==0.5.0
ai21==4.0.3
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
configobj==5.0.9
configparser==7.2.0
dataclasses-json==0.6.7
datasets==3.6.0
decorator==5.2.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
einops==0.8.1
etelemetry==0.3.1
evaluate==0.4.3
exceptiongroup==1.3.0
executing==2.2.0
faiss-cpu==1.11.0
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.3.0
greenlet==3.2.3
h11==0.16.0
hf-xet==1.1.4
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
idna==3.10
importlib_resources==6.5.2
ipython==8.37.0
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-openai==0.3.24
langchain-text-splitters==0.3.8
langsmith==0.3.45
looseversion==1.3.0
lxml==5.4.0
markdown2==2.5.3
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.5.0
multiprocess==0.70.16
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai==1.88.0
orjson==3.10.18
packaging==24.2
pandas==2.3.0
parso==0.8.4
pathlib==1.0.1
pexpect==4.9.0
pillow==11.2.1
prompt_toolkit==3.0.51
propcache==0.3.2
prov==2.0.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.29
pyarrow==20.0.0
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydot==4.0.1
Pygments==2.19.1
PyMuPDF==1.26.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
ragas==0.2.15
rdflib==7.1.4
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
sentence-transformers==4.1.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
stack-data==0.6.3
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.1
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
transformers==4.52.4
triton==3.3.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
